# Optimized std.datetime.systime.SysTime.fromISOString