So this rendition is a bit messier than normal because of the issue of using `find`. `indexOf` doesn't have a multiple `needle`s, so I tried to refactor to use `countUntil`, but it ended up being worse readability wise because you need to know which of the elements was found. 

The `static if`s are there because the return type of `byCodeUnit` is different if the source is an array of `dchar`s. In hindsight, this doesn't seem like this was the best idea but w/e.

No performance numbers for LDC because it doesn't have the change which gives access to `source` yet

```
$ dmd -O -inline -release test.d && ./test
2013-Feb-07 04:39:37.0000503
old		2 secs, 316 ms, 927 μs, and 3 hnsecs
new		1 sec, 350 ms, 355 μs, and 4 hnsecs
```

<details>

```d
import std.stdio;
import std.algorithm;
import std.exception;
import std.conv;
import std.ascii;
import std.range;
import std.traits;
import std.typecons;
import std.string;
import std.datetime.systime;
import std.datetime.date;
import std.datetime.timezone;
import std.datetime.stopwatch;
import core.time;
import std.utf;
import std.random;

enum testCount = 1_000_000;

__gshared a = "20130207T043937.000050392";

SysTime fromISOString1(S)(in S isoString, immutable TimeZone tz = null) @safe
        if (isSomeString!S)
{
    import std.algorithm.searching : startsWith, find;
    import std.conv : to;
    import std.string : strip;

    auto dstr = to!dstring(strip(isoString));
    immutable skipFirst = dstr.startsWith('+', '-') != 0;

    auto found = (skipFirst ? dstr[1..$] : dstr).find('.', 'Z', '+', '-');
    auto dateTimeStr = dstr[0 .. $ - found[0].length];

    dstring fracSecStr;
    dstring zoneStr;

    if (found[1] != 0)
    {
        if (found[1] == 1)
        {
            auto foundTZ = found[0].find('Z', '+', '-');

            if (foundTZ[1] != 0)
            {
                fracSecStr = found[0][0 .. $ - foundTZ[0].length];
                zoneStr = foundTZ[0];
            }
            else
                fracSecStr = found[0];
        }
        else
            zoneStr = found[0];
    }

    try
    {
        auto dateTime = DateTime.fromISOString(dateTimeStr);
        auto fracSec = fracSecsFromISOString(fracSecStr);
        Rebindable!(immutable TimeZone) parsedZone;

        if (zoneStr.empty)
            parsedZone = LocalTime();
        else if (zoneStr == "Z")
            parsedZone = UTC();
        else
        {
            try
                parsedZone = fromISOString(zoneStr);
            catch (DateTimeException dte)
                parsedZone = fromISOExtString(zoneStr);
        }

        auto retval = SysTime(dateTime, fracSec, parsedZone);

        if (tz !is null)
            retval.timezone = tz;

        return retval;
    }
    catch (DateTimeException dte)
        throw new DateTimeException(format("Invalid ISO String: %s", isoString));
}

SysTime fromISOString2(S)(in S isoString, immutable TimeZone tz = null) @safe
        if (isSomeString!S)
{
    import std.algorithm.searching : startsWith, find;
    import std.conv : to;
    import std.string : strip;
    import std.utf : byCodeUnit;

    auto str = strip(isoString);
    immutable skipFirst = str.startsWith('+', '-');

    auto found = (skipFirst ? str[1..$] : str).byCodeUnit.find('.', 'Z', '+', '-');
    auto dateTimeStr = str[0 .. $ - found[0].length];

    typeof(str) fracSecStr;
    typeof(str) zoneStr;

    if (found[1] != 0)
    {
        if (found[1] == 1)
        {
            auto foundTZ = found[0].find('Z', '+', '-');

            if (foundTZ[1] != 0)
            {
                static if (isNarrowString!S)
                {
                    fracSecStr = found[0][0 .. $ - foundTZ[0].length].source; 
                    zoneStr = foundTZ[0].source;
                }
                else
                {
                    fracSecStr = found[0][0 .. $ - foundTZ[0].length]; 
                    zoneStr = foundTZ[0];
                }
            }
            else
            {
                static if (isNarrowString!S)
                    fracSecStr = found[0].source;
                else
                    fracSecStr = found[0];
            }
        }
        else
        {
            static if (isNarrowString!S)
                zoneStr = found[0].source;
            else
                zoneStr = found[0];
        }
    }

    try
    {
        auto dateTime = DateTime.fromISOString(dateTimeStr);
        auto fracSec = fracSecsFromISOString(fracSecStr);

        Rebindable!(immutable TimeZone) parsedZone;

        if (zoneStr.empty)
            parsedZone = LocalTime();
        else if (zoneStr == "Z")
            parsedZone = UTC();
        else
        {
            try
                parsedZone = fromISOString(zoneStr);
            catch (DateTimeException dte)
                parsedZone = fromISOExtString(zoneStr);
        }

        auto retval = SysTime(dateTime, fracSec, parsedZone);

        if (tz !is null)
            retval.timezone = tz;

        return retval;
    }
    catch (DateTimeException dte)
        throw new DateTimeException(format("Invalid ISO String: %s", isoString));
}

Duration fracSecsFromISOString(S)(in S isoString) @trusted pure
if (isSomeString!S)
{
    import std.algorithm.searching : all;
    import std.ascii : isDigit;
    import std.conv : to;
    import std.string : representation;
    import core.time;

    if (isoString.empty)
        return Duration.zero;

    auto str = isoString.representation;

    enforce(str[0] == '.', new DateTimeException("Invalid ISO String"));
    str.popFront();

    enforce(!str.empty && all!isDigit(str), new DateTimeException("Invalid ISO String"));

    dchar[7] fullISOString = void;
    foreach (i, ref dchar c; fullISOString)
    {
        if (i < str.length)
            c = str[i];
        else
            c = '0';
    }

    return hnsecs(to!int(fullISOString[]));
}

immutable(SimpleTimeZone) fromISOString(S)(S isoString) @safe pure
        if (isSomeString!S)
{
    import std.algorithm.searching : startsWith;
    import std.conv : text, to, ConvException;
    import std.datetime.date : DateTimeException;
    import std.exception : enforce;

    auto whichSign = isoString.startsWith('-', '+');
    enforce!DateTimeException(whichSign > 0, text("Invalid ISO String ", isoString));

    isoString = isoString[1 .. $];
    auto sign = whichSign == 1 ? -1 : 1;
    int hours;
    int minutes;

    try
    {
        // cast to int from uint is used because it checks for
        // non digits without extra loops
        if (isoString.length == 2)
        {
            hours = cast(int) to!uint(isoString);
        }
        else if (isoString.length == 4)
        {
            hours = cast(int) to!uint(isoString[0 .. 2]);
            minutes = cast(int) to!uint(isoString[2 .. 4]);
        }
        else
        {
            throw new DateTimeException(text("Invalid ISO String ", isoString));
        }
    }
    catch (ConvException)
    {
        throw new DateTimeException(text("Invalid ISO String ", isoString));
    }

    enforce!DateTimeException(hours < 24 && minutes < 60, text("Invalid ISO String ", isoString));

    return new immutable SimpleTimeZone(sign * (dur!"hours"(hours) + dur!"minutes"(minutes)));
}

immutable(SimpleTimeZone) fromISOExtString(S)(S isoExtString) @safe pure
        if (isSomeString!S)
{
    import std.algorithm.searching : startsWith;
    import std.conv : ConvException, to;
    import std.datetime.date : DateTimeException;
    import std.exception : enforce;
    import std.format : format;
    import std.string : indexOf;

    auto whichSign = isoExtString.startsWith('-', '+');
    enforce!DateTimeException(whichSign > 0, format("Invalid ISO String: %s", isoExtString));
    auto sign = whichSign == 1 ? -1 : 1;

    isoExtString = isoExtString[1 .. $];
    enforce!DateTimeException(!isoExtString.empty, format("Invalid ISO String: %s", isoExtString));

    immutable colon = isoExtString.indexOf(':');
    S hoursStr;
    S minutesStr;
    int hours, minutes;

    if (colon != -1)
    {
        hoursStr = isoExtString[0 .. colon];
        minutesStr = isoExtString[colon + 1 .. $];
        enforce!DateTimeException(minutesStr.length == 2, format("Invalid ISO String: %s", isoExtString));
    }
    else
    {
        hoursStr = isoExtString;
    }

    enforce!DateTimeException(hoursStr.length == 2, format("Invalid ISO String: %s", isoExtString));

    try
    {
        // cast to int from uint is used because it checks for
        // non digits without extra loops
        hours = cast(int) to!uint(hoursStr);
        minutes = cast(int) (minutesStr.empty ? 0 : to!uint(minutesStr));
    }
    catch (ConvException)
    {
        throw new DateTimeException(format("Invalid ISO String: %s", isoExtString));
    }

    enforce!DateTimeException(hours < 24 && minutes < 60, format("Invalid ISO String: %s", isoExtString));

    return new immutable SimpleTimeZone(sign * (dur!"hours"(hours) + dur!"minutes"(minutes)));
}

void main()
{
    SysTime res;

    auto result = to!Duration(benchmark!(() => res = fromISOString1(a))(testCount)[0]);
    auto result2 = to!Duration(benchmark!(() => res = fromISOString2(a))(testCount)[0]);

    writeln(res); // force execution

    writeln("old", "\t\t", result);
    writeln("new", "\t\t", result2);
}
```

</details>