@JackStouffer 2018-02-07
@wilzbach The trouble is that `byCodeUnit` on a `dstring` returns `dstring`.