@wilzbach 2018-02-07
@JackStouffer this is what I'm talking about:

```diff
diff --git a/std/datetime/systime.d b/std/datetime/systime.d
index 52f777e95..3676db1f2 100644
--- a/std/datetime/systime.d
+++ b/std/datetime/systime.d
@@ -8105,8 +8105,8 @@ public:
         auto found = (skipFirst ? str[1..$] : str).byCodeUnit.find('.', 'Z', '+', '-');
         auto dateTimeStr = str[0 .. $ - found[0].length];
 
-        typeof(str) fracSecStr;
-        typeof(str) zoneStr;
+        typeof(found[0].source) fracSecStr;
+        typeof(found[0].source) zoneStr;
 
         if (found[1] != 0)
         {
@@ -8116,31 +8116,17 @@ public:
 
                 if (foundTZ[1] != 0)
                 {
-                    static if (isNarrowString!S)
-                    {
-                        fracSecStr = found[0][0 .. $ - foundTZ[0].length].source;
-                        zoneStr = foundTZ[0].source;
-                    }
-                    else
-                    {
-                        fracSecStr = found[0][0 .. $ - foundTZ[0].length];
-                        zoneStr = foundTZ[0];
-                    }
+                    fracSecStr = found[0].source[0 .. $ - foundTZ[0].length];
+                    zoneStr = foundTZ[0].source;
                 }
                 else
                 {
-                    static if (isNarrowString!S)
-                        fracSecStr = found[0].source;
-                    else
-                        fracSecStr = found[0];
+                    fracSecStr = found[0].source;
                 }
             }
             else
             {
-                static if (isNarrowString!S)
-                    zoneStr = found[0].source;
-                else
-                    zoneStr = found[0];
+                zoneStr = found[0].source;
             }
         }
 
diff --git a/std/utf.d b/std/utf.d
index 71681d368..58e67d43a 100644
--- a/std/utf.d
+++ b/std/utf.d
@@ -3345,12 +3345,22 @@ if (isAutodecodableString!R ||
     else static if (is(R : const dchar[]) && !__traits(hasMember, R, "empty") &&
                     !__traits(hasMember, R, "front") && !__traits(hasMember, R, "popFront"))
     {
-        return cast(StringTypeOf!R) r;
+        static struct ByCodeUnitImpl
+        {
+            StringTypeOf!R source;
+            alias source this;
+        }
+        return ByCodeUnitImpl(cast(StringTypeOf!R) r);
     }
     else
     {
         // byCodeUnit for ranges and dchar[] is a no-op
-        return r;
+        static struct ByCodeUnitImpl
+        {
+            R source;
+            alias source this;
+        }
+        return ByCodeUnitImpl(r);
     }
 }
```

but I currently don't have time to dig into why it's making the `std.utf` testsuite fail / whether we can actually do this. It does pass your PR tests though.