@JackStouffer 2018-02-07
I don't think we should focus on the `byCodeUnit` problem at the moment because the problem of changing the return value in this way is something I don't think we should do so lightly.