@jmdavis 2018-02-07
It would break tons of code if `byCodeUnit` suddenly returned a wrapper struct for dstrings. Sure, your stuck with more `static if`s in order to strip of `byCodeUnitImpl` from narrow strings, but it works. Until a release or two back, not even that was possible, because it didn't have `source`. But in order to make it so that `source` worked for anything passed to `byCodeUnit`, you'd have to wrap all types, and that would break existing code. It's the same problem when you use `take` or `takeExactly` on different types of ranges where some support slicing and some don't.