@jmdavis 2018-02-07
My solution to the static ifs in one of my projects was to create a `stripBCU` function which returned the original range if it wasn't wrapped in a `ByCodeUnitImpl` and returned `source` if it was, but because `ByCodeUnitImpl` is a Voldemort type, you have to pass it the original type as a template argument so that it knows whether to strip off `ByCodeUnitImpl` (it can't look for `source`, since the original range might have a `source`). So, maybe we should break `ByCodeUnitImpl` out as `ByCodeUnit` so that it can be introspected (maybe even providing `stripBCU` in std.utf), but that's a separate issue. The static ifs work just fine. They're just a bit noisy, unfortunately.